refactor: rename to keyple-interop-localreader-nfcmobile
