+++ gradle.properties
@@ -1,7 +1,7 @@
-title = Keyple Kotlin Multiplatform NFC Reader Library
-description = Keyple NFC Reader KMP Library - A Kotlin Multiplatform implementation enabling NFC card communications across Android, iOS and desktop platforms
+title = Keyple Kotlin Multiplatform NFC Mobile Local Reader Library
+description = Keyple KMP NFC Mobile Local Reader Library - A Kotlin Multiplatform implementation enabling NFC card communications across Android, iOS and desktop platforms
 group = org.eclipse.keyple
-version = 0.1.5
+version = 0.1.6
 
 kotlin.code.style=official
