+++ src/iosMain/kotlin/org/eclipse/keyple/interop/localreader/nfcmobile/NfcReader.kt
@@ -11,7 +11,7 @@
  ************************************************************************************** */
 @file:OptIn(ExperimentalStdlibApi::class, BetaInteropApi::class)
 
-package org.eclipse.keyple.keypleless.reader.nfcmobile
+package org.eclipse.keyple.interop.localreader.nfcmobile
 
 import io.github.aakira.napier.Napier
 import kotlin.coroutines.resume
@@ -29,8 +29,8 @@ import kotlinx.coroutines.MainScope
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
-import org.eclipse.keyple.keypleless.distributed.client.spi.CardIOException
-import org.eclipse.keyple.keypleless.distributed.client.spi.ReaderIOException
+import org.eclipse.keyple.interop.jsonapi.client.spi.CardIOException
+import org.eclipse.keyple.interop.jsonapi.client.spi.ReaderIOException
 import platform.CoreNFC.NFCISO7816APDU
 import platform.CoreNFC.NFCISO7816TagProtocol
 import platform.CoreNFC.NFCPollingISO14443
